requirements.txt
asgiref==3.5.2
Django==4.1.2
django-wkhtmltopdf==3.4.0
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.3
